Add editUser page
